Modified 'Free' graphic, updated generator code to reflect changes
The 'Free' graphic was made into it's own separate file to be overlayed whenever the free tile is activated in the board spec configuration settings. In the code, the variable 'overlay' was appropriately refactored to be 'foreground' and 'free_overlay' is now the variable containing the 'Free' text graphic which gets overlayed on top of everything else if so configured.